requirements.txt
# Function dependencies, for example:
# package>=version
# Function dependencies, for example:
# package>=version
Flask
gunicorn
google-api-core
google-cloud-bigquery
pandas
pandas-gbq
gcsfs
bs4
